+++ packages/expo-modules-autolinking/src/platforms/devtools.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 
 import type { ExtraDependencies, ModuleDescriptorDevTools, PackageRevision } from '../types';
+import { isPathInside, maybeRealpath } from '../utils';
 
 export async function resolveModuleAsync(
   packageName: string,
@@ -14,8 +15,21 @@ export async function resolveModuleAsync(
   return {
     packageName,
     packageRoot: revision.path,
-    webpageRoot: path.join(revision.path, devtoolsConfig.webpageRoot),
+    webpageRoot: await resolveWebpageRoot(revision.path, devtoolsConfig.webpageRoot),
   };
+}
+
+async function resolveWebpageRoot(
+  packageRoot: string,
+  configuredWebpageRoot: string | undefined
+): Promise<string | undefined> {
+  if (!configuredWebpageRoot) {
+    return undefined;
+  }
+  const resolvedWebpageRoot = path.resolve(packageRoot, configuredWebpageRoot);
+  // NOTE(@kitten): Failing realpath-ing, typically due to ENOENT, results in the original value
+  const webpageRoot = (await maybeRealpath(resolvedWebpageRoot)) ?? resolvedWebpageRoot;
+  return isPathInside(webpageRoot, packageRoot) ? webpageRoot : undefined;
 }
 
 export async function resolveExtraBuildDependenciesAsync(